refactor(app): extract the update wiring
The update reminder, the in-app swap-and-relaunch flow, the manual
"Check now" button and the once-a-day background check move to
`wire::update`. Bodies unchanged.
